@@ -12,42 +12,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/cli": "^1.7.4",
-    "@angular/common": "^5.2.9",
-    "@angular/compiler": "^5.2.9",
-    "@angular/compiler-cli": "^5.2.9",
-    "@angular/core": "^5.2.9",
-    "@angular/forms": "^5.2.9",
-    "@angular/http": "^5.2.9",
-    "@angular/platform-browser": "^5.2.9",
-    "@angular/platform-browser-dynamic": "^5.2.9",
-    "@angular/platform-server": "^5.2.9",
-    "@angular/router": "^5.2.9",
     "core-js": "^2.5.5",
-    "font-awesome": "^4.7.0",
-    "rxjs": "^5.5.9",
-    "webpack": "^3.11.0",
-    "zone.js": "^0.8.26"
+    "font-awesome": "^4.7.0"
   },
   "devDependencies": {
-    "@angular/language-service": "^5.2.9",
+    "@angular-devkit/build-angular": "~0.6.3",
+    "@angular/animations": "^6.0.2",
+    "@angular/cli": "^6.0.3",
+    "@angular/common": "^6.0.2",
+    "@angular/compiler": "^6.0.2",
+    "@angular/compiler-cli": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "@angular/forms": "^6.0.2",
+    "@angular/http": "^6.0.2",
+    "@angular/language-service": "^6.0.2",
+    "@angular/platform-browser": "^6.0.2",
+    "@angular/platform-browser-dynamic": "^6.0.2",
+    "@angular/platform-server": "^6.0.2",
+    "@angular/router": "^6.0.2",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.105",
     "codelyzer": "~3.0.1",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
+    "karma": "^2.0.2",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.4.2",
     "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "primeng": "^5.2.4",
-    "protractor": "~5.1.2",
+    "protractor": "^5.3.2",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "^2.6.2"
+    "typescript": "^2.7.2",
+    "webpack": "^4.8.3",
+    "zone.js": "^0.8.26"
   }
 }
